requirements.txt
aiogram==3.2.0
asyncpg==0.29.0
redis==5.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
